Merged PR 3: Add initial backend structure
I took the project structure from [this](https://github.com/tiangolo/full-stack-fastapi-postgresql) code generator and threw out everything we don't need.

This relies heavily on data models from `neomodel` and `pydantic`. Not sure if we should work with the `neo4j` driver directly, seems like more work to me, so I went with `neomodel`.

This contains a working example of registering a user with the DB, including tests.

The structures and names of the `APIRouter`s might not be ideal yet, I guess we can change those later, but before we integrate it with the backend.

I'm not very happy with the way Neo4j is integrated here. I tried to follow the style of the generated project, but `neomodel` doesn't seem to have the concept of sessions, so it's very different from PostgreSQL.

Authentication is also missing for now and I think it's more urgent than I thought, because things like creating a post should imho get the author's identity from the auth token and not from an API parameter. Maybe we should use PostgreSQL (or any other RDBMS supported by SQLAlchemy) for the user information? In that case we could use the boilerplate from the code generator for authentication. Having 3 instead of 2 databases would add complexity that can be avoided, so I'm not sure what the best solution is here.

Related work items: #47