Merged PR 32: Refactor posts & comments
I think this refactoring is useful, especially down the road. Since posts and comments have some similarities (like timestamps, author, upvotes, etc.), I expanded the concept of a common `Thing` superclass to reduce duplicate code (e.g. setting/getting the author) and avoid unnecessary case distinctions (e.g. parent of a comment).

I already had this in the Neo4j model, but it's now being used throughout the application. It relies a lot on `Generic`s and the `Comment` schema might look a little bloated (having the `text` nested inside `content`), but imho this is well worth it to be able to use `Thing` for upvotes, parent of comments, etc.

This might be a pain to review if you're not that familiar with [generics](https://www.python.org/dev/peps/pep-0484/#generics) in Python. I didn't add any new features and all tests pass, so if you agree with the concept of using `Thing`s where possible, feel free to just skim over this. I'm pretty confident this doesn't break anything, as we have a pretty good test coverage.

As I'd like to base the upvote logic on this refactoring, I'd be happy about a timely review.

Related work items: #101