Merged PR 36: Change get_posts to only return posts from one subreddit
This is a quick and dirty fix for the endpoint discussed earlier. Basically, I think we have to offer at least these three possibilities to get posts:

- get posts for the frontpage for a user that is logged in (showing posts from all of his subreddits); this can only be implemented once @<Merle Hoffmann> finishes her work
- get posts for the frontpage for a user that is **not** logged in (showing posts from all subreddits); this is what the function was returning before
- get posts for a specific subreddit (showing only posts from that subreddit); this is what the function now does, after my commit

I decided to not implement both of the two last options yet. I want to wait until I can implement the first option as well, as I hope to offer the first two options in one endpoint.

In the meantime, the last option is the most important one I guess, so here it is.

Related work items: #107